--- a/requirements-ext.txt
+++ b/requirements-ext.txt
@@ -1,4 +1,3 @@
 numpy
 singledispatch
 six
-PuLP>=1.6.5