This PR implements our own IP for coffee mode instead of calling PuLP

- Passes all firedrake tests.
- Matches the flop counts on my test suits.
- Time performance is largely the same as now, if not slightly better (e.g. holzapfel from 5.2s to 4.7s, hyperelasticity in 3D from 3.6s to 3.3s).

Inefficiencies in the algorithm:

- Can do better by keeping track of violated constraints, instead of repeated testing constraints that are already fulfilled across levels.

This supersedes, and thus closes #117.