@blechta 2017-06-29
This is great! The FFC regression tests (or the output in general) will not depend on `glpsol` version.