@wence- 2017-07-21
Since @tj-sun is away for much of the rest of the summer, I might have a go at cleaning this up next week as part of py3 (in part to remove our compiled dependency list).