@wence- 2017-07-24
OK, this produces same flop counts as before my changes, so will merge.